fix(safari menu issue): fixed the menu issue in safari (#758)
* fix(safari menu issue): fixed the menu issue in safari

- added Product and Resources to array object for both mobile and desktop view
- added order to change the ordering in mobile view
- minor JSX clean ups

* chore: variable name fix for readablity

* feat(safari fix): navigation structure href key to URL to support mobile subtopic structure

* refactor(topnav): improve dropdown item rendering

Refactor the rendering of dropdown items in the TopNav component to improve readability and accessibility. Instead of using the "href" attribute directly, use the "Link" component from Next.js for better URL handling. Also, update the structure of the dropdown item to include the item name and description in separate elements, improving the visual presentation. This change enhances the user experience and ensures consistent rendering across different devices and browsers.